mediascan/backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
flask-jwt-extended==4.6.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.3
python-dotenv==1.0.1
Pillow==10.4.0
numpy==1.26.4
reportlab==4.2.2
gunicorn==22.0.0
psycopg2-binary==2.9.9
tensorflow==2.16.1
opencv-python-headless==4.9.0.80
torch==2.3.0
torchvision==0.18.0
pandas==2.2.2
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
requests==2.31.0
supabase==2.4.3
gotrue==2.3.0
